Moved fetchEvents and subscription porcelain to a single search porcelain section
