+++ src/social/content.ts
@@ -4,8 +4,6 @@
 import { ActivityPubCreateOpts } from "../activityPub/activityPub";
 import { Config } from "../config/config";
 import { Handle } from "./handles";
-import { DSNPMessage } from "../types/DSNP";
-import { KeccakHash } from "../utilities/hash";
 
 /**
  * broadcast() creates a broadcast DSNP event and enqueues it for the next
@@ -68,36 +66,4 @@ export const react = async (activityPubOpts: ActivityPubCreateOpts, opts: Config
   //
   // const event = events.createReactionEvent(uri, inReplyTo, activityPubHash);
   // config.batchEnqueueMethod(event, opts);
-};
-
-/**
- * @typedef EventFilters
- * @property {string[]} types          The type of events to fetch, i.e. "notes", "follows" etc.
- * @property {Handle[]} includeHandles List of user handles to fetch events from. If not present, all users not listed in excludeHandles will be included.
- * @property {Handle[]} excludeHandles List of user handles to exclude from results, such as a blocklist.
- * @property {KeccakHash[]} inReplyTo  List of hashes to seek replies to. For use in fetching comments and reactions.
- * @property {string} limit            The maximum number of results to fetch.
- * @property {BlockNumber} to          The last block to scan for events. Defaults to current block.
- * @property {BlockNumber} from        The first block to scan for events. Defaults to genesis.
- */
-interface EventFilters {
-  types: string[];
-  includeHandles?: Handle[];
-  excludeHandles?: Handle[];
-  inReplyTo?: KeccakHash[];
-  limit?: number;
-  to?: BlockNumber;
-  from?: BlockNumber;
-}
-
-/**
- * fetchEvents() fetches the most recent activity pub events matching the given
- * search criteria. This method is not yet implemented.
- *
- * @param filters  Any filter options for including or excluding certain events
- * @param opts     Optional. Configuration overrides, such as from address, if any
- * @returns        An array of events
- */
-export const fetchEvents = async (filters: EventFilters): DSNPMessage[] => {
-  throw NotImplementedError();
 };
